mpi/topo: separate some more complicated topo functions
Now that mpi-level source code are in binding/c/, we can move some the
more complex functions from topo_impl.c to its own source file for
better organization.
