+++ src/mpi/topo/Makefile.mk
@@ -4,5 +4,7 @@
 ##
 
 mpi_core_sources +=       \
+    src/mpi/topo/dims_create.c \
+    src/mpi/topo/dist_graph_create.c \
     src/mpi/topo/topo_impl.c \
     src/mpi/topo/topoutil.c
